requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mssql-django==1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
pyodbc==4.0.39
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wfastcgi==3.0.0
whitenoise==6.5.0
yarl==1.9.2
